Revert "Hyermdev"

+++ src/main/kotlin/com/example/cakezip/domain/cake/CakeOptionList.kt
@@ -20,7 +20,7 @@ class CakeOptionList(
 
     var optionDetail: String,
 
-    var optionPrice: Int,
+    var optionPrice: Long,
 
     ) : BaseEntity() {
     @Id
